Refactoring mass matrix package
