+++ xfv/src/mass_matrix/one_dimension_mass_matrix.py
@@ -1,5 +1,5 @@
-#!/usr/bin/env python2.7
-# -*- coding: iso-8859-1 -*-
+#!/usr/bin/env python3.7
+# -*- coding: utf-8 -*-
 """
 Implementing the OneDimensionMassMatrix class
 """
